+++ SurvivalUtility.java
@@ -1,7 +1,7 @@
 public class SurvivalUtility extends Utility {
 
     @Override
-    public int evaluate(TetrisState predictedState) {
+    public double evaluate(TetrisState predictedState) {
         if(predictedState.hasLost()) {
             return 0;
         }
